(svn r5892) -Fix (r5887): silence MSVC warning(s) about signedness of some comparisons with _date.

+++ openttd.h
@@ -62,7 +62,7 @@ typedef uint32 WindowNumber;
 typedef byte WindowClass;
 
 typedef uint8  Year;
-typedef uint32 Date;
+typedef uint16 Date;
 
 
 enum GameModes {
